Cleaned up the code and some comments. Separated some classes into their own files.
