--- FSEJobFinder-ConsoleApp/FSEDataFeed/ConsoleMain.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace FSEDataFeed
-{
-    class ConsoleMain
-    {
-        static void main(string[] args)
-        {
-           
-        }
-    }
-}
