+++ FSEJobFinder-ConsoleApp/FSEDataFeed/FSEDataExport.cs
@@ -1,11 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using System.Net;
-using System.Text;
-using System.Threading.Tasks;
-using System.Xml;
 using System.Xml.Serialization;
 
 
@@ -22,7 +18,6 @@ namespace FSEDataFeed
     /// </summary>
     public class FSEDataExport  
     {   
-
         private string userKey;
         private int accessKeyHitCount;
         private string FSEEndpoint;
